requirements.txt
appnope==0.1.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.4.1
asttokens==2.0.5
attrs==19.2.0
backcall==0.1.0
beautifulsoup4==4.8.1
black==21.12b0
bleach==3.1.0
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.10
click==8.0.3
cloudpickle==2.0.0
croniter==1.0.15
cycler==0.10.0
dask==2021.11.1
decorator==4.4.0
defusedxml==0.6.0
distributed==2021.11.1
dj-database-url==0.5.0
Django==3.2.8
django-boost==1.7.2
django-cleanup==5.2.0
django-cors-headers==3.10.1
django-environ==0.8.1
django-filter==21.1
django-heroku==0.0.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
docker==5.0.3
entrypoints==0.3
et-xmlfile==1.1.0
executing==0.8.2
fsspec==2021.11.0
gunicorn==20.1.0
HeapDict==1.0.1
homebrew==0.1.10
idna==2.8
importlib-metadata==4.8.2
importlib-resources==5.4.0
injector==0.18.4
ipykernel==5.1.2
ipython==7.8.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
Janome==0.3.9
jedi==0.15.1
Jinja2==2.10.3
jsonschema==3.0.2
jupyter==1.0.0
jupyter-client==5.3.3
jupyter-console==6.0.0
jupyter-core==4.5.0
jupyterlab-pygments==0.1.2
kiwisolver==1.1.0
locket==0.2.1
lxml==4.4.1
MarkupSafe==1.1.1
marshmallow==3.14.0
marshmallow-oneofschema==3.0.1
matplotlib-inline==0.1.3
mistune==0.8.4
msgpack==1.0.2
mypy-extensions==0.4.3
mysql-connector-python==8.0.27
Naked==0.1.31
nbclient==0.5.10
nbconvert==5.6.0
nbformat==4.4.0
nest-asyncio==1.5.4
notebook==6.0.1
openpyxl==3.0.9
packaging==21.2
pandocfilters==1.4.2
parso==0.5.1
partd==1.2.0
pathspec==0.9.0
pendulum==2.1.2
pexpect==4.7.0
pickleshare==0.7.5
platformdirs==2.4.1
prefect==0.15.9
prometheus-client==0.7.1
prompt-toolkit==2.0.10
protobuf==3.19.1
psutil==5.8.0
ptyprocess==0.6.0
pure-eval==0.2.1
pycparser==2.21
pycryptodome==3.15.0
Pygments==2.4.2
PyJWT==1.7.1
PyMySQL==1.0.2
pyparsing==2.4.2
pyrsistent==0.15.4
python-box==5.4.1
python-dateutil==2.8.0
python-slugify==5.0.2
pytz==2019.3
pytzdata==2020.1
PyYAML==6.0
pyzmq==18.1.0
qrcode==6.1
qtconsole==4.5.5
QtPy==2.0.0
requests==2.22.0
schedule==0.6.0
Send2Trash==1.5.0
shellescape==3.8.1
six==1.12.0
sortedcontainers==2.4.0
soupsieve==1.9.4
sqlparse==0.4.2
stack-data==0.1.4
tabulate==0.8.9
tblib==1.7.0
terminado==0.8.2
testpath==0.4.2
text-unidecode==1.3
toml==0.10.2
tomli==1.2.3
toolz==0.11.2
tornado==6.0.3
traitlets==4.3.3
typing-extensions==3.10.0.2
ua-parser==0.10.0
urllib3==1.25.6
user-agents==2.2.0
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==1.2.1
whitenoise==5.3.0
widgetsnbextension==3.5.1
zict==2.0.0
zipp==3.6.0
